requirements.txt
streamlit
requests
beautifulsoup4
fpdf
feedparser
google-generativeai